requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-deps -o requirements.txt pyproject.toml
flock-core==0.5.0b3
    # via flock-showcase (pyproject.toml)
ipykernel==6.29.5
    # via flock-showcase (pyproject.toml)
ipywidgets==8.1.7
    # via flock-showcase (pyproject.toml)
jupyter==1.1.1
    # via flock-showcase (pyproject.toml)
